+++ lib/testcentricity_web/web_core/page_objects_helper.rb
@@ -662,6 +662,8 @@ module TestCentricity
             actual = ui_object.count(visible = :all)
           when :siebel_options
             actual = ui_object.get_siebel_options
+          when :style
+            actual = ui_object.style
           when :aria_label
             actual = ui_object.aria_label
           when :aria_disabled